Address second round of review feedback on #3968
- Rename the transform parameter from trsf to transformation for
  readability, per review.
- Drop the redundant ap["/N"].get_object() call:
  DictionaryObject.__getitem__ already resolves indirect references
  and returns a plain PdfObject (never None), so the None guard
  added for the previous round was genuinely dead code, exactly as
  flagged. Removing the redundant resolution removes the need for
  the guard entirely rather than working around the type checker.
- Move the six annotation-appearance tests from test_page.py to
  test_appearance_stream.py, matching where the function they test
  now lives.
