+++ tests/test_appearance_stream.py
@@ -7,9 +7,17 @@ from unittest import mock
 
 import pytest
 
-from pypdf import PdfWriter
+from pypdf import PdfWriter, Transformation
 from pypdf._font import Font
-from pypdf.generic import RectangleObject
+from pypdf.generic import (
+    ArrayObject,
+    DecodedStreamObject,
+    DictionaryObject,
+    FloatObject,
+    NameObject,
+    NumberObject,
+    RectangleObject,
+)
 from pypdf.generic._appearance_stream import BaseStreamAppearance, BaseStreamConfig, TextStreamAppearance
 
 from . import RESOURCE_ROOT
@@ -236,3 +244,193 @@ def test_base_stream_config_rotation(rotation, outcome):
     )
     appearance = BaseStreamAppearance(layout=layout)
     assert appearance.get("/Matrix", None) == outcome
+
+
+def test_merge_transformed_page_annotation_with_ap_but_no_normal_state():
+    """/AP present but missing /N (malformed, but must not raise) is a no-op
+    for the appearance transform -- /Rect still moves normally.
+    """
+    src_writer = PdfWriter()
+    src_page = src_writer.add_blank_page(width=200, height=200)
+    annotation = DictionaryObject()
+    annotation[NameObject("/Type")] = NameObject("/Annot")
+    annotation[NameObject("/Subtype")] = NameObject("/Square")
+    annotation[NameObject("/Rect")] = RectangleObject((20, 20, 120, 70))
+    annotation[NameObject("/AP")] = DictionaryObject()  # no /N key
+    src_writer.add_annotation(0, annotation)
+
+    dest_writer = PdfWriter()
+    dest_page = dest_writer.add_blank_page(width=200, height=200)
+    dest_page.merge_transformed_page(src_page, Transformation().translate(10, 5))
+
+    merged = dest_page["/Annots"][0].get_object()
+    assert tuple(round(float(x), 6) for x in merged["/Rect"]) == (30.0, 25.0, 130.0, 75.0)
+
+
+def test_merge_transformed_page_annotation_with_malformed_normal_state():
+    """/N resolving to something other than a stream or a state dict (a
+    malformed PDF) must be skipped, not raise.
+    """
+    src_writer = PdfWriter()
+    src_page = src_writer.add_blank_page(width=200, height=200)
+    annotation = DictionaryObject()
+    annotation[NameObject("/Type")] = NameObject("/Annot")
+    annotation[NameObject("/Subtype")] = NameObject("/Square")
+    annotation[NameObject("/Rect")] = RectangleObject((20, 20, 120, 70))
+    annotation[NameObject("/AP")] = DictionaryObject({NameObject("/N"): NumberObject(5)})
+    src_writer.add_annotation(0, annotation)
+
+    dest_writer = PdfWriter()
+    dest_page = dest_writer.add_blank_page(width=200, height=200)
+    dest_page.merge_transformed_page(src_page, Transformation().translate(10, 5))
+
+    merged = dest_page["/Annots"][0].get_object()
+    assert tuple(round(float(x), 6) for x in merged["/Rect"]) == (30.0, 25.0, 130.0, 75.0)
+    assert merged["/AP"]["/N"] == 5
+
+
+def test_merge_transformed_page_updates_annotation_appearance_matrix():
+    """
+    An annotation's /Rect is repositioned/resized by the merge transform, but
+    per the appearance-stream algorithm (PDF 2.0, 12.5.5) a viewer fits the
+    appearance's /BBox into /Rect with an axis-aligned scale only, never a
+    rotation. Leaving /AP /N's /Matrix untouched under a rotating transform
+    therefore stretches the original, unrotated appearance content into the
+    new, differently-shaped /Rect instead of rotating it. The transform must
+    be composed into the appearance's own /Matrix as well.
+    """
+    src_writer = PdfWriter()
+    src_page = src_writer.add_blank_page(width=200, height=200)
+
+    ap_stream = DecodedStreamObject()
+    ap_stream.set_data(b"0 0 m 100 50 l S")
+    ap_stream[NameObject("/Type")] = NameObject("/XObject")
+    ap_stream[NameObject("/Subtype")] = NameObject("/Form")
+    ap_stream[NameObject("/BBox")] = ArrayObject(
+        [FloatObject(0), FloatObject(0), FloatObject(100), FloatObject(50)]
+    )
+    ap_ref = src_writer._add_object(ap_stream)
+
+    annotation = DictionaryObject()
+    annotation[NameObject("/Type")] = NameObject("/Annot")
+    annotation[NameObject("/Subtype")] = NameObject("/Square")
+    annotation[NameObject("/Rect")] = RectangleObject((20, 20, 120, 70))
+    annotation[NameObject("/AP")] = DictionaryObject({NameObject("/N"): ap_ref})
+    src_writer.add_annotation(0, annotation)
+
+    dest_writer = PdfWriter()
+    dest_page = dest_writer.add_blank_page(width=200, height=200)
+    transform = Transformation().rotate(90).translate(200, 0)
+    dest_page.merge_transformed_page(src_page, transform)
+
+    merged_ap = dest_page["/Annots"][0].get_object()["/AP"]["/N"].get_object()
+    # A pure 90 degree rotation + translate composed onto an identity
+    # starting matrix: (0, 1, -1, 0, 200, 0).
+    matrix = tuple(round(float(x), 6) for x in merged_ap["/Matrix"])
+    assert matrix == (0.0, 1.0, -1.0, 0.0, 200.0, 0.0)
+    # The appearance stream's own local geometry (/BBox) is untouched; only
+    # /Matrix carries the transform, consistent with how content streams and
+    # /Rect are already handled elsewhere in this function.
+    assert tuple(merged_ap["/BBox"]) == (0, 0, 100, 50)
+
+
+def test_merge_transformed_page_composes_existing_annotation_matrix():
+    """An annotation that already has its own /Matrix must have the merge
+    transform composed on top of it, not overwrite it outright.
+    """
+    src_writer = PdfWriter()
+    src_page = src_writer.add_blank_page(width=200, height=200)
+
+    ap_stream = DecodedStreamObject()
+    ap_stream.set_data(b"0 0 m 100 50 l S")
+    ap_stream[NameObject("/Type")] = NameObject("/XObject")
+    ap_stream[NameObject("/Subtype")] = NameObject("/Form")
+    ap_stream[NameObject("/BBox")] = ArrayObject(
+        [FloatObject(0), FloatObject(0), FloatObject(100), FloatObject(50)]
+    )
+    # Pre-existing 2x horizontal scale.
+    ap_stream[NameObject("/Matrix")] = ArrayObject(
+        [FloatObject(v) for v in (2, 0, 0, 1, 0, 0)]
+    )
+    ap_ref = src_writer._add_object(ap_stream)
+
+    annotation = DictionaryObject()
+    annotation[NameObject("/Type")] = NameObject("/Annot")
+    annotation[NameObject("/Subtype")] = NameObject("/Square")
+    annotation[NameObject("/Rect")] = RectangleObject((0, 0, 200, 50))
+    annotation[NameObject("/AP")] = DictionaryObject({NameObject("/N"): ap_ref})
+    src_writer.add_annotation(0, annotation)
+
+    dest_writer = PdfWriter()
+    dest_page = dest_writer.add_blank_page(width=200, height=200)
+    # Translate-only this time, to isolate composition order from rotation.
+    transform = Transformation().translate(10, 5)
+    dest_page.merge_transformed_page(src_page, transform)
+
+    merged_ap = dest_page["/Annots"][0].get_object()["/AP"]["/N"].get_object()
+    matrix = tuple(round(float(x), 6) for x in merged_ap["/Matrix"])
+    # (2, 0, 0, 1, 0, 0) composed with a (10, 5) translation: the scale is
+    # preserved and the translation is appended, not scaled by it.
+    assert matrix == (2.0, 0.0, 0.0, 1.0, 10.0, 5.0)
+
+
+def test_merge_transformed_page_annotation_without_appearance_stream():
+    """Annotations with no /AP at all (the common case) must merge exactly
+    as before -- only /Rect moves, and nothing raises.
+    """
+    src_writer = PdfWriter()
+    src_page = src_writer.add_blank_page(width=200, height=200)
+    annotation = DictionaryObject()
+    annotation[NameObject("/Type")] = NameObject("/Annot")
+    annotation[NameObject("/Subtype")] = NameObject("/Square")
+    annotation[NameObject("/Rect")] = RectangleObject((20, 20, 120, 70))
+    src_writer.add_annotation(0, annotation)
+
+    dest_writer = PdfWriter()
+    dest_page = dest_writer.add_blank_page(width=200, height=200)
+    transform = Transformation().rotate(90).translate(200, 0)
+    dest_page.merge_transformed_page(src_page, transform)
+
+    merged = dest_page["/Annots"][0].get_object()
+    assert "/AP" not in merged
+    assert tuple(round(float(x), 6) for x in merged["/Rect"]) == (130.0, 20.0, 180.0, 120.0)
+
+
+def test_merge_transformed_page_annotation_with_multi_state_appearance():
+    """Widget annotations with multiple states (e.g. checkboxes) store /AP /N
+    as a dict of named sub-streams rather than a single stream. Every state's
+    /Matrix must be updated, not just the first, and nothing should raise.
+    """
+    src_writer = PdfWriter()
+    src_page = src_writer.add_blank_page(width=200, height=200)
+
+    states = {}
+    for name in ("/Off", "/Yes"):
+        state_stream = DecodedStreamObject()
+        state_stream.set_data(b"0 0 m 100 50 l S")
+        state_stream[NameObject("/Type")] = NameObject("/XObject")
+        state_stream[NameObject("/Subtype")] = NameObject("/Form")
+        state_stream[NameObject("/BBox")] = ArrayObject(
+            [FloatObject(0), FloatObject(0), FloatObject(100), FloatObject(50)]
+        )
+        states[NameObject(name)] = src_writer._add_object(state_stream)
+
+    annotation = DictionaryObject()
+    annotation[NameObject("/Type")] = NameObject("/Annot")
+    annotation[NameObject("/Subtype")] = NameObject("/Widget")
+    annotation[NameObject("/Rect")] = RectangleObject((20, 20, 120, 70))
+    annotation[NameObject("/AP")] = DictionaryObject(
+        {NameObject("/N"): DictionaryObject(states)}
+    )
+    src_writer.add_annotation(0, annotation)
+
+    dest_writer = PdfWriter()
+    dest_page = dest_writer.add_blank_page(width=200, height=200)
+    transform = Transformation().rotate(90).translate(200, 0)
+    dest_page.merge_transformed_page(src_page, transform)
+
+    merged_states = dest_page["/Annots"][0].get_object()["/AP"]["/N"]
+    assert set(merged_states.keys()) == {"/Off", "/Yes"}
+    for state in merged_states.values():
+        matrix = tuple(round(float(x), 6) for x in state.get_object()["/Matrix"])
+        assert matrix == (0.0, 1.0, -1.0, 0.0, 200.0, 0.0)
